Separate global config from service config
ConfigurationServicePlaceholders references all clients, this means when a single client is deep imported it still transitively references all clients.

The impact of this is TypeScript needs to parse all type definitions in the SDK, and there is no way for a project to avoid this.

By separating client config from global config clients can be deep imported without pulling all type definitions in, greatly reducing build time
