+++ lib/credentials/process_credentials.d.ts
@@ -1,5 +1,5 @@
 import {Credentials} from '../credentials';
-import {HTTPOptions} from '../config';
+import {HTTPOptions} from '../config-base';
 export class ProcessCredentials extends Credentials {
     /**
      * Creates a new ProcessCredentials object.
