+++ lib/credentials/shared_ini_file_credentials.d.ts
@@ -1,5 +1,5 @@
 import {Credentials} from '../credentials';
-import {HTTPOptions} from '../config';
+import {HTTPOptions} from '../config-base';
 export class SharedIniFileCredentials extends Credentials {
     /**
      * Creates a new SharedIniFileCredentials object.
